separate concerns of account and database creds into separate classes
