+++ src/commands/key.mjs
@@ -5,23 +5,15 @@ import { container } from "../cli.mjs";
 async function createKey(argv) {
   const logger = container.resolve("logger");
   const AccountClient = container.resolve("accountClient");
-  // const accountKey = await getAccountKey(profile);
-  // TODO: after logging in, should we list the top level databases and create db keys for them?
-  //  depending on how many top level dbs....
-  // Have to list DBs on login so we know which databases are top-level and require frontdoor calls
-
-  // TODO: we should create the key with fauna unless it's a top level key
-  // in which case we should create it with the account client
-
-  // TODO: when using fauna to create a key at the specified database path, we should
-  //  getDBKey(parent path).
-  // const dbSecret = getDBKey({
-  //   accountKey,
-  //   path: database,
-  //   role,
-  // });
-  // logger.stdout("got account key", accountKey);
-  // logger.stdout("got db secret", dbSecret);
+  const { database, role, ttl } = argv;
+  const databaseKey = await AccountClient.createKey({
+    path: database,
+    role,
+    ttl,
+  });
+  logger.stdout(
+    `Created key for ${database} with role ${role}\n${JSON.stringify(databaseKey)}`,
+  );
 }
 
 function buildKeyCommand(yargs) {
