+++ src/config/setup-container.mjs
@@ -13,7 +13,8 @@ import updateNotifier from "update-notifier";
 import { parseYargs } from "../cli.mjs";
 import { performV4Query, performV10Query } from "../commands/eval.mjs";
 import { makeAccountRequest } from "../lib/account.mjs";
-import { Credentials } from "../lib/auth/credentials.mjs";
+import { AccountCreds } from "../lib/auth/accountCreds.mjs";
+import { DatabaseCreds } from "../lib/auth/databaseCreds.mjs";
 import OAuthClient from "../lib/auth/oauth-client.mjs";
 import { getSimpleClient } from "../lib/command-helpers.mjs";
 import { makeFaunaRequest } from "../lib/db.mjs";
@@ -74,10 +75,13 @@ export const injectables = {
 
   // While we inject the class instance before this in middleware,
   //  we need to register it here to resolve types.
-  credentials: awilix.asClass(Credentials, {
+  accountClient: awilix.asClass(FaunaAccountClient, {
     lifetime: Lifetime.SINGLETON,
   }),
-  accountClient: awilix.asClass(FaunaAccountClient, {
+  accountCreds: awilix.asClass(AccountCreds, {
+    lifetime: Lifetime.SINGLETON,
+  }),
+  databaseCreds: awilix.asClass(DatabaseCreds, {
     lifetime: Lifetime.SINGLETON,
   }),
